+++ src/app/(navbar)/search/page.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ForwardBox from "./_component/ForwardBox";
-import ContentsBox from "./_component/ContentsBox";
+import ContentsBox from "../../_component/common/layout/ContentsBox";
 import SearchHeader from "./_component/SearchHeader";
 import HashtagsBox from "../../_component/common/layout/HashtagsBox";
