+++ src/app/_component/common/layout/ContentsBox.tsx
@@ -8,7 +8,7 @@ interface Props {
 const ContentsBox = ({ title, children }: Props) => {
   return (
     <section className="px-6">
-      <p className="my-6 text-black-2 font-semibold text-[18px] whitespace-pre-line">
+      <p className="my-6 text-black-2 font-semibold text-lg whitespace-pre-line">
         {title}
       </p>
       {children}
